Add env var override to all configuration settings
This allows users to configure VSS-Server entirely through environment
variables, without a configuration file. We hence remove the requirement
that a path to a configuration file always be passed as an
argument.

If a file is passed as an argument, and it does not exist,
we now abort startup.

Also consolidate host and port settings into single socket address
setting.